Remove GribFile's iteration capability
This is accompanied by some smaller changes to the behavior in GribMessage.
Specifically, I discovered that if I tested the arguments used to instantiate a
GribMessage, it would result in the file descriptor returned by the GRIB API
being invalid. Also, iterating directly over all messages in a GRIB file using
the GribFile class returned the same message over and over again. For some
reason, wrapping too much complexity around getting a new message from a GRIB
file causes the GRIB API to return a GID without an error, but this GID will
always correspond to the first message in the file. The file seeks to the end of
the first message and remains there.

Consequently, I have removed the ability to iterate directly over GRIB messages
and have reimplemented the initialization method of GribMessage. Tests and
documentation have been modified accordingly.